+++ test/unit/mysql_log_admin/main.py
@@ -28,6 +28,142 @@ import version
 __version__ = version.__version__
 
 
+class ArgParser(object):
+
+    """Class:  ArgParser
+
+    Description:  Class stub holder for gen_class.ArgParser class.
+
+    Methods:
+        __init__
+        arg_cond_req
+        arg_dir_chk
+        arg_require
+        arg_validate
+        arg_xor_dict
+        get_args
+        get_val
+
+    """
+
+    def __init__(self):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+
+        """
+
+        self.cmdline = None
+        self.args_array = dict()
+        self.opt_con_req = None
+        self.opt_con_req2 = True
+        self.dir_perms_chk = None
+        self.dir_perms_chk2 = True
+        self.opt_req = None
+        self.opt_req2 = True
+        self.opt_xor_val = None
+        self.opt_xor_val2 = True
+        self.valid_func = None
+        self.valid_func2 = True
+
+    def arg_cond_req(self, opt_con_req):
+
+        """Method:  arg_cond_req
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_cond_req.
+
+        Arguments:
+
+        """
+
+        self.opt_con_req = opt_con_req
+
+        return self.opt_con_req2
+
+    def arg_dir_chk(self, dir_perms_chk):
+
+        """Method:  arg_dir_chk
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_dir_chk.
+
+        Arguments:
+
+        """
+
+        self.dir_perms_chk = dir_perms_chk
+
+        return self.dir_perms_chk2
+
+    def arg_require(self, opt_req):
+
+        """Method:  arg_require
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_require.
+
+        Arguments:
+
+        """
+
+        self.opt_req = opt_req
+
+        return self.opt_req2
+
+    def arg_validate(self, valid_func):
+
+        """Method:  arg_validate
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_validate.
+
+        Arguments:
+
+        """
+
+        self.valid_func = valid_func
+
+        return self.valid_func2
+
+    def arg_xor_dict(self, opt_xor_val):
+
+        """Method:  arg_xor_dict
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_xor_dict.
+
+        Arguments:
+
+        """
+
+        self.opt_xor_val = opt_xor_val
+
+        return self.opt_xor_val2
+
+    def get_args(self):
+
+        """Method:  get_args
+
+        Description:  Method stub holder for gen_class.ArgParser.get_args.
+
+        Arguments:
+
+        """
+
+        return self.args_array
+
+    def get_val(self, skey, def_val):
+
+        """Method:  get_val
+
+        Description:  Method stub holder for gen_class.ArgParser.get_val.
+
+        Arguments:
+
+        """
+
+        return self.args_array.get(skey, def_val)
+
+
 class ProgramLock(object):
 
     """Class:  ProgramLock
@@ -65,22 +201,20 @@ class UnitTest(unittest.TestCase):
         setUp
         test_help_true
         test_help_false
-        test_arg_req_true
         test_arg_req_false
         test_arg_req_true
-        test_arg_req_false
         test_arg_xor_false
         test_arg_xor_true
-        test_arg_dir_true
         test_arg_dir_false
+        test_arg_dir_true
         test_arg_validate_false
         test_arg_validate_true
         test_arg_cond_req_false
         test_arg_cond_req_true
         test_run_program
-        test_programlock_id
-        test_programlock_false
         test_programlock_true
+        test_programlock_false
+        test_programlock_id
 
     """
 
@@ -94,13 +228,13 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.args_array = {"-c": "CfgFile", "-d": "CfgDir"}
-        self.args_array2 = {"-c": "CfgFile", "-d": "CfgDir", "-y": "Flavor"}
         self.proglock = ProgramLock(["cmdline"], "FlavorID")
+        self.args = ArgParser()
 
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser.arg_parse2")
-    def test_help_true(self, mock_arg, mock_help):
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_help_true(self, mock_arg):
 
         """Function:  test_help_true
 
@@ -110,14 +244,14 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.return_value = self.args_array
-        mock_help.return_value = True
+        mock_arg.return_value = self.args
 
         self.assertFalse(mysql_log_admin.main())
 
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_help_false(self, mock_arg, mock_help):
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_help_false(self, mock_arg):
 
         """Function:  test_help_false
 
@@ -127,33 +261,16 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = True
+        self.args.opt_req2 = False
+
+        mock_arg.return_value = self.args
 
         self.assertFalse(mysql_log_admin.main())
 
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_arg_req_true(self, mock_arg, mock_help):
-
-        """Function:  test_arg_req_true
-
-        Description:  Test arg_require if returns true.
-
-        Arguments:
-
-        """
-
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = True
-
-        self.assertFalse(mysql_log_admin.main())
-
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_arg_req_false(self, mock_arg, mock_help):
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_arg_req_false(self, mock_arg):
 
         """Function:  test_arg_req_false
 
@@ -163,16 +280,35 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_xor_dict.return_value = False
+        self.args.opt_req2 = False
+
+        mock_arg.return_value = self.args
 
         self.assertFalse(mysql_log_admin.main())
 
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_arg_xor_false(self, mock_arg, mock_help):
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_arg_req_true(self, mock_arg):
+
+        """Function:  test_arg_req_true
+
+        Description:  Test arg_require if returns true.
+
+        Arguments:
+
+        """
+
+        self.args.opt_xor_val2 = False
+
+        mock_arg.return_value = self.args
+
+        self.assertFalse(mysql_log_admin.main())
+
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_arg_xor_false(self, mock_arg):
 
         """Function:  test_arg_xor_false
 
@@ -182,16 +318,16 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_xor_dict.return_value = False
+        self.args.opt_xor_val2 = False
+
+        mock_arg.return_value = self.args
 
         self.assertFalse(mysql_log_admin.main())
 
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_arg_xor_true(self, mock_arg, mock_help):
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_arg_xor_true(self, mock_arg):
 
         """Function:  test_arg_xor_true
 
@@ -201,38 +337,16 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = True
+        self.args.dir_perms_chk2 = False
+
+        mock_arg.return_value = self.args
 
         self.assertFalse(mysql_log_admin.main())
 
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_arg_dir_true(self, mock_arg, mock_help):
-
-        """Function:  test_arg_dir_true
-
-        Description:  Test arg_dir_chk_crt if returns true.
-
-        Arguments:
-
-        """
-
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_valid_val.return_value = True
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = True
-
-        self.assertFalse(mysql_log_admin.main())
-
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_arg_dir_false(self, mock_arg, mock_help):
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_arg_dir_false(self, mock_arg):
 
         """Function:  test_arg_dir_false
 
@@ -242,19 +356,35 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_valid_val.return_value = True
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
-        mock_arg.arg_validate.return_value = False
+        self.args.dir_perms_chk2 = False
+
+        mock_arg.return_value = self.args
 
         self.assertFalse(mysql_log_admin.main())
 
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_arg_validate_false(self, mock_arg, mock_help):
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_arg_dir_true(self, mock_arg):
+
+        """Function:  test_arg_dir_true
+
+        Description:  Test arg_dir_chk_crt if returns true.
+
+        Arguments:
+
+        """
+
+        self.args.valid_func2 = False
+
+        mock_arg.return_value = self.args
+
+        self.assertFalse(mysql_log_admin.main())
+
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_arg_validate_false(self, mock_arg):
 
         """Function:  test_arg_validate_false
 
@@ -264,19 +394,16 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_valid_val.return_value = True
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
-        mock_arg.arg_validate.return_value = False
+        self.args.valid_func2 = False
+
+        mock_arg.return_value = self.args
 
         self.assertFalse(mysql_log_admin.main())
 
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_arg_validate_true(self, mock_arg, mock_help):
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_arg_validate_true(self, mock_arg):
 
         """Function:  test_arg_validate_true
 
@@ -286,20 +413,16 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_valid_val.return_value = True
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
-        mock_arg.arg_validate.return_value = True
-        mock_arg.arg_cond_req.return_value = False
+        self.args.opt_con_req2 = False
+
+        mock_arg.return_value = self.args
 
         self.assertFalse(mysql_log_admin.main())
 
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_arg_cond_req_false(self, mock_arg, mock_help):
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_arg_cond_req_false(self, mock_arg):
 
         """Function:  test_cond_req_false
 
@@ -309,22 +432,18 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_valid_val.return_value = True
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
-        mock_arg.arg_validate.return_value = True
-        mock_arg.arg_cond_req.return_value = False
+        self.args.opt_con_req2 = False
+
+        mock_arg.return_value = self.args
 
         self.assertFalse(mysql_log_admin.main())
 
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.run_program", mock.Mock(return_value=True))
     @mock.patch("mysql_log_admin.gen_class.ProgramLock")
-    @mock.patch("mysql_log_admin.run_program")
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_arg_cond_req_true(self, mock_arg, mock_help, mock_run, mock_lock):
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_arg_cond_req_true(self, mock_arg, mock_lock):
 
         """Function:  test_arg_cond_req_true
 
@@ -334,24 +453,17 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_valid_val.return_value = True
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
-        mock_arg.arg_validate.return_value = True
-        mock_arg.arg_cond_req.return_value = True
-        mock_run.return_value = True
+        mock_arg.return_value = self.args
         mock_lock.return_value = self.proglock
 
         self.assertFalse(mysql_log_admin.main())
 
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.run_program", mock.Mock(return_value=True))
     @mock.patch("mysql_log_admin.gen_class.ProgramLock")
-    @mock.patch("mysql_log_admin.run_program")
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_run_program(self, mock_arg, mock_help, mock_run, mock_lock):
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_run_program(self, mock_arg, mock_lock):
 
         """Function:  test_run_program
 
@@ -361,78 +473,17 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_valid_val.return_value = True
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
-        mock_arg.arg_validate.return_value = True
-        mock_arg.arg_cond_req.return_value = True
-        mock_run.return_value = True
+        mock_arg.return_value = self.args
         mock_lock.return_value = self.proglock
 
         self.assertFalse(mysql_log_admin.main())
 
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.run_program", mock.Mock(return_value=True))
     @mock.patch("mysql_log_admin.gen_class.ProgramLock")
-    @mock.patch("mysql_log_admin.run_program")
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_programlock_id(self, mock_arg, mock_help, mock_run, mock_lock):
-
-        """Function:  test_programlock_id
-
-        Description:  Test ProgramLock fails with flavor id.
-
-        Arguments:
-
-        """
-
-        mock_arg.arg_parse2.return_value = self.args_array2
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_valid_val.return_value = True
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
-        mock_arg.arg_validate.return_value = True
-        mock_arg.arg_cond_req.return_value = True
-        mock_run.return_value = True
-        mock_lock.return_value = self.proglock
-
-        self.assertFalse(mysql_log_admin.main())
-
-    @mock.patch("mysql_log_admin.gen_class.ProgramLock")
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_programlock_false(self, mock_arg, mock_help, mock_lock):
-
-        """Function:  test_programlock_false
-
-        Description:  Test with ProgramLock returns False.
-
-        Arguments:
-
-        """
-
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_valid_val.return_value = True
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
-        mock_arg.arg_validate.return_value = True
-        mock_arg.arg_cond_req.return_value = True
-        mock_lock.side_effect = \
-            mysql_log_admin.gen_class.SingleInstanceException
-
-        with gen_libs.no_std_out():
-            self.assertFalse(mysql_log_admin.main())
-
-    @mock.patch("mysql_log_admin.gen_class.ProgramLock")
-    @mock.patch("mysql_log_admin.run_program")
-    @mock.patch("mysql_log_admin.gen_libs.help_func")
-    @mock.patch("mysql_log_admin.arg_parser")
-    def test_programlock_true(self, mock_arg, mock_help, mock_run, mock_lock):
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_programlock_true(self, mock_arg, mock_lock):
 
         """Function:  test_programlock_true
 
@@ -442,15 +493,50 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args_array
-        mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_valid_val.return_value = True
-        mock_arg.arg_xor_dict.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
-        mock_arg.arg_validate.return_value = True
-        mock_arg.arg_cond_req.return_value = True
-        mock_run.return_value = True
+        mock_arg.return_value = self.args
+        mock_lock.return_value = self.proglock
+
+        self.assertFalse(mysql_log_admin.main())
+
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ProgramLock")
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_programlock_false(self, mock_arg, mock_lock):
+
+        """Function:  test_programlock_false
+
+        Description:  Test with ProgramLock returns False.
+
+        Arguments:
+
+        """
+
+        mock_arg.return_value = self.args
+        mock_lock.side_effect = \
+            mysql_log_admin.gen_class.SingleInstanceException
+
+        with gen_libs.no_std_out():
+            self.assertFalse(mysql_log_admin.main())
+
+    @mock.patch("mysql_log_admin.gen_libs.help_func",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.run_program", mock.Mock(return_value=True))
+    @mock.patch("mysql_log_admin.gen_class.ProgramLock")
+    @mock.patch("mysql_log_admin.gen_class.ArgParser")
+    def test_programlock_id(self, mock_arg, mock_lock):
+
+        """Function:  test_programlock_id
+
+        Description:  Test ProgramLock fails with flavor id.
+
+        Arguments:
+
+        """
+
+        self.args.args_array = {"-y": "flavor"}
+
+        mock_arg.return_value = self.args
         mock_lock.return_value = self.proglock
 
         self.assertFalse(mysql_log_admin.main())