+++ CHANGELOG.md
@@ -7,7 +7,10 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ## [3.0.0] - 2023-02-24
 Breaking Changes
 
-- Set up instance of the gen_class.ArgParser class which replaces all manual operations of the args_array variable that contain the command line arguments.
+- Replaced args_parser module with the gen_class.ArgParser class.
+
+### Changed
+- main: Replaced the use of arg_parser (args_array) with gen_class.ArgParser class (args).
 
 
 ## [2.2.2] - 2022-10-12